+++ tests/ExtraParagraphTest.php
@@ -14,7 +14,7 @@ final class ExtraParagraphTest extends TestCase
     $html = $formatter->Format($document);
 
     $this->assertEquals(
-      '<p><span style="font-weight:bold;font-family:Calibri;font-size:16px;color:#000000;">Conditions<br/></span><span style="font-family:Calibri;font-size:16px;color:#000000;">&#1;Delivery: FCA in our warehouse in Rotterdam<br/>&#1;Lead Time: 25 working days after confirmation, subject to prior sale<br/>&#1;Payment: 60 days after invoice date<br/>&#1;Quote validity: 30 days',
+      '<p><span style="font-weight:bold;font-family:Calibri;font-size:16px;color:#000000;">Conditions<br/></span><span style="font-family:Calibri;font-size:16px;color:#000000;">Delivery: FCA in our warehouse in Rotterdam<br/>Lead Time: 25 working days after confirmation, subject to prior sale<br/>Payment: 60 days after invoice date<br/>Quote validity: 30 days</span></p>',
       $html
     );
   }